// src/shared/api/generated/operation/listImportRunsParams.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Amazing EKB Hub API
 * MVP backend API (OpenAPI-first)
 * OpenAPI spec version: 1.0.0
 */
import type { ImportRunStatus } from '../model/importRunStatus';

export type ListImportRunsParams = {
/**
 * Фильтр по content source.
 */
sourceId?: string;
/**
 * Фильтр по статусу import run.
 */
status?: ImportRunStatus;
};